Extract the prompt composer from chat
kanban's inline workspace form is a second prompt composer: it reached
into chat for the attachment preview, both palettes and four chat-input
hooks -- seven of the fourteen remaining barrel-bypassing imports, all
from one file. Re-exporting those through chat's barrel would have made
the barrel a dumping ground for internals with one outside caller each.

None of the seven modules knows anything about a chat session. Their
whole dependency set is @/lib, @/components/ui, @/client/lib/trpc and
each other, so they lift out whole into src/client/features/composer/
and both callers consume the same barrel. cubic suggested src/client/lib
for the four hooks; that is where non-component utilities live, and three
of these are React components, so they would not have fit.

The palette regression tests stay in chat and import through the new
barrel -- splitting them would have duplicated 45 lines of jsdom harness
to separate three describe blocks. The inline form's seven vi.mock calls
collapse into one factory covering all seven names, since a partial
factory leaves the rest undefined rather than falling through.

14 barrel-bypassing cross-feature imports drop to 7.
