+++ src/client/features/composer/attachment-preview.tsx
