+++ src/client/features/composer/file-mention-palette.tsx
