+++ src/client/features/composer/hooks/use-paste-drop-handler.ts
