+++ src/client/features/composer/palette-keyboard-navigation.ts
