+++ src/client/features/composer/slash-command-palette.tsx
